+++ Assets/Scripts/Economy/UpgradesFunction.cs
@@ -18,7 +18,7 @@ public class UpgradesFunction : MonoBehaviour
     {
         if(pfMain._score >= 15)
         {
-            pfMain._score = EconomyMain.deductScore(15, pfMain._score);
+            pfMain._score = EconomyMain.deductScore(15, (float)pfMain._score);
             pfMain._scoreToAdd += 1;
         }
     }
@@ -27,7 +27,7 @@ public class UpgradesFunction : MonoBehaviour
     {
         if (pfMain._score >= 50)
         {
-            pfMain._score = EconomyMain.deductScore(50, pfMain._score);
+            pfMain._score = EconomyMain.deductScore(50, (float)pfMain._score);
             pfMain._passiveScore += 1;
         }
     }
@@ -36,7 +36,7 @@ public class UpgradesFunction : MonoBehaviour
     {
         if (pfMain._score >= 60)
         {
-            pfMain._score = EconomyMain.deductScore(60, pfMain._score);
+            pfMain._score = EconomyMain.deductScore(60, (float)pfMain._score);
             pfMain._passiveScore += 2;
         }
     }
@@ -45,7 +45,7 @@ public class UpgradesFunction : MonoBehaviour
     {
         if (pfMain._score >= 65)
         {
-            pfMain._score = EconomyMain.deductScore(65, pfMain._score);
+            pfMain._score = EconomyMain.deductScore(65, (float)pfMain._score);
             pfMain._passiveScore += 5;
         }
     }
@@ -54,7 +54,7 @@ public class UpgradesFunction : MonoBehaviour
     {
         if (pfMain._score >= 100)
         {
-            pfMain._score = EconomyMain.deductScore(100, pfMain._score);
+            pfMain._score = EconomyMain.deductScore(100, (float)pfMain._score);
             pfMain._passiveScore += 10;
         }
     }
@@ -63,7 +63,7 @@ public class UpgradesFunction : MonoBehaviour
     {
         if (pfMain._score >= 120)
         {
-            pfMain._score = EconomyMain.deductScore(120, pfMain._score);
+            pfMain._score = EconomyMain.deductScore(120, (float)pfMain._score);
             pfMain._passiveScore += 20;
         }
     }
@@ -72,7 +72,7 @@ public class UpgradesFunction : MonoBehaviour
     {
         if (pfMain._score >= 150)
         {
-            pfMain._score = EconomyMain.deductScore(150, pfMain._score);
+            pfMain._score = EconomyMain.deductScore(150, (float)pfMain._score);
             pfMain._passiveScore += 30;
         }
     }
@@ -81,7 +81,7 @@ public class UpgradesFunction : MonoBehaviour
     {
         if (pfMain._score >= 200)
         {
-            pfMain._score = EconomyMain.deductScore(200, pfMain._score);
+            pfMain._score = EconomyMain.deductScore(200, (float)pfMain._score);
             pfMain._passiveScore += 50;
         }
     }
@@ -90,7 +90,7 @@ public class UpgradesFunction : MonoBehaviour
     {
         if (pfMain._score >= 1000)
         {
-            pfMain._score = EconomyMain.deductScore(1000, pfMain._score);
+            pfMain._score = EconomyMain.deductScore(1000, (float)pfMain._score);
             pfMain._passiveScore += 100;
         }
     }
